backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
sentence-transformers
faiss-cpu
numpy
pandas
google-generativeai
email-validator
bcrypt==4.0.1
